[ospfd] Do not fall through to the intervening router case.
If the second case in 16.1.1, para 5 has been found, always
return to caller even if no nexthop was found. It is incorrect
to treat a missing next hop as it is intervening router
around the corner.

Remove misplaced test for unnumberd links too, no PtP links
can be connected to a Network vertex.